Convert to ES6.
Now an ES6 implementation. Many changes but similar in spirit to the
original. Cache is the base object. The sub-caches are now infinitely
nestable for what that's worth. Also means that you can have no nested
caches at all, which is how I'm using the cache I wrote in Strata to
replace Magazine there. (Now Strata will use this Magazine.)

Going to make content cache a wrapper object with the auto magic
expected of an NPM LRU built into it there.

Wrote an extensive Docco based readme and took a pass at the
`README.md`. Was looking to finish it before commit, but feel like I
should take a break and commit and release since the documentation is
suggesting changes to support content caching. Also, wouldn't hurt to
alpha release in order to try to apply this new cache in Strata and the
R-Tree to see if it works as well as expected.

Closes #97.