Rethinking eviction iterator.
Should be a synchronous thing. The iteration functions can be more
generally documented as acting on entry properties that they do not
control, that they can only advise on.

As far as external iteration for eviction, that has a lot of caveats.
The `least()` interface is probably better because it looks and acts
similar to `hold()`, you have to think about race conditions and
therefore you have to think about locking. At some point I wanted
external eviction because I thought needed to do something async, but
ultimately that async activity took place in a queue and release and
eviction was synchronous again.

See #97.