Create sub-caches without providing key.
Someone may recreate a sub-cache with the key. This is the case if
someone closes and reopens a database and uses the database name as the
key. They are going to have to anticipate this and include something
unique so that the the reopened database does not read the cache entries
of the closed database. A problem if the user is just counting on
eviction to get rid of the old pages.