Implement an async cache helper.
This is not based on `heft`, just on the count. Could do heft probably.
Not even sure if this belongs in this library or if it should be a
part of Strata where it will be used. Wanted to unit test it apart from
having to simulate errors with file handles, but it would be easy enough
to leave it an generic class, but keep it specific to Strata. Then I
don't have to document it here, nor to I have to bulk up the Cache class
with all these overlays.